+++ packages/react-querybuilder/src/types/basic.ts
@@ -125,7 +125,7 @@ interface BaseFullField<
   ValueObj extends FullOption = FullOption<ValueName>,
 > extends WithOptionalClassName<BaseFullOption<FieldName>> {
   id?: string;
-  operators?: FlexibleOptionList<OperatorObj>;
+  operators?: FlexibleOptionList<OperatorObj> | OperatorName[] | FlexibleOption<OperatorName>[] | (OperatorName | FlexibleOption<OperatorName>)[];
   valueEditorType?: ValueEditorType | ((operator: OperatorName) => ValueEditorType);
   valueSources?:
     | ValueSources